refactor theme sass logic into its own module
in preperation to remove this module from Open edX in Tahoe 2.0
