+++ openedx/core/djangoapps/appsembler/theme_sass/tests/test_helpers.py
@@ -1,0 +1,3 @@
+"""
+Tests for the theme_sass module helpers.
+"""
